Seperated query functions into seperate files
