+++ query_time.go
@@ -1,0 +1,1 @@
+package rethinkgo
